Upgraded to Axum 0.7
Using newest version of Axum (0.7.5) and Hyper which is beyond 1.0 bringing more stability. Since Hyper doesn’t include a client, we are using the one suggested in Hyper-Utils, which is a drop-in replacement from the client that was present in Hyper 0.14.

Hyper now relies on Http-Body 1 which cascaded into a bunch of changes to HTTP bodies used throughout the code. Hyper also gives the option to pick any execution framework but we’re sticking to Tokio.