Allow PRSS to generate a single value in one call
This closes #1029 by adding a method to PRSS implementation that allows callers to specify which randomness they want to obtain: shared with the helper on the right or on the left. This saves 50% of computation in cases where only one side of PRSS is required. #1029 lists the cases where it is useful.