Improve pre-commit runtime by running fewer checks
Our observation it takes 10+ minutes for all tests to pass and it is blocking people from working on other things. This change intends to keep only basic checks where we see the majority of failures: formatting, clippy and basic tests.

It is still possible to run all checks by setting  `EXEC_SLOW_TESTS` environment variable to 1