Add a handler to kill a query
In our testing with external parties, it became quite evident that
bouncing helpers if query fails is far from having good experience
for people who run the test. We need to have an ability to kill queries
that failed on one or more helper and restart the runs.

This change does exactly that - adding a `/query_id/kill` handler
that will respond with immediately aborting the task that runs
an IPA query.