Parametrize MAC-malicious contexts with ShardBinding parameter
This is just a first step to enable sharded circuits to be written with malicious security. It is far from being complete, but it gradually moves us towards that goal.

This change just enables malicious contexts to support sharding. Because of type aliases, everything else from the outside of this module still uses non-sharded versions