Support sharded malicious protocols inside test infrastructure
This is the final PR in the series (#1333, #1315) that enables writing sharded malicious protocols.

The support is added for `Runner::malicious` function, and it can be used to execute malicious circuits. Other methods like `upgraded_semi_honest` and `upgraded_malicious` haven't been touched yet - I am not convinced that these are necessary and we can update them as we go.

The core goal here is to unblock efforts to implement Hybrid protocol