Allow customizing channel size between collector and producer thread
I noticed in unit tests it gets very hard to get a correct snapshot after test finishes. Often test thread sends the data and moves on, but the collector thread hasn't processed the data yet. So any assertion on the metric fails occasionally.

One way to deal with it is to block the sender until receiver processes its data. Obviously performance-wise it is not great, but it is ok for unit tests