Support cross-shard PRSS in sharded MPC circuits
We ran into an issue where shards need coordinated randomness in more than one protocol:

* Malicious shuffle needs it to generate MAC keys
* PRF needs it to generate PRF key to mask the values

At this point I strongly suspect we will need it in more places and the current mitigation
is not sustainable. We do simulate key exchange in shuffle and about to in PRF, so I decided
to stop the bleeding and implemented coordinated randomness for PRSS.

The only API difference between those two is to use cross-shard PRSS, one would call
`ShardedContext::cross_shard_prss` and shard-scoped PRSS is still available by calling
`ShardedContext::prss`. Both return the same structure, so anything that can be generated with
shard-scoped PRSS, can also be generated with cross-shard randomness. It is basically a drop-in
replacement, only available for sharded MPC circuits.

Note that this only brings in-memory implementation, but I have implemented 50% of distribution
circuit already and it looks pretty straightforward - just generate a key using PRSS and distribute
it across shards - it will be used to bootstrap generators on every shard. This will be a separate
PR and will take some time to implement - but folks shouldn't be blocked as this change unblocks
using the in-memory infrastructure.