Use stream buffering in report collector (#1504)
* Use stream buffering in report collector

Continuation of #1501, we want to avoid submitting reports one by one from RC to each individual shard. That likely leads to fragmentation and we've been observing slow execution on the client side.

This sets up the buffer size to be divisible by TCP MSS, but I don't have any real evidence that this is going to work well. We would need to experiment with it

* Use 8Kb buffers for streaming inside report collector